Move tests into gitserver/integration_tests
