+++ cmd/frontend/backend/mocks.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/sourcegraph/sourcegraph/internal/actor"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver/integration_tests"
 	"github.com/sourcegraph/sourcegraph/internal/trace/ot"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 )
 
 var Mocks MockServices
@@ -17,7 +17,7 @@ type MockServices struct {
 // testContext creates a new context.Context for use by tests
 func testContext() context.Context {
 	Mocks = MockServices{}
-	git.ResetMocks()
+	integration_tests.ResetMocks()
 
 	ctx := context.Background()
 	ctx = actor.WithActor(ctx, &actor.Actor{UID: 1})
