+++ internal/codeintel/uploads/background/cleanup/unknown_commits_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/shared"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver/integration_tests"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 )
 
 func init() {
@@ -103,7 +103,7 @@ func testUnknownCommitsJanitor(t *testing.T, resolveRevisionFunc func(commit str
 	gitserver.Mocks.ResolveRevision = func(spec string, opt gitserver.ResolveRevisionOptions) (api.CommitID, error) {
 		return api.CommitID(spec), resolveRevisionFunc(spec)
 	}
-	defer git.ResetMocks()
+	defer integration_tests.ResetMocks()
 
 	dbStore := NewMockDBStore()
 	dbStore.TransactFunc.SetDefaultReturn(dbStore, nil)
