+++ internal/gitserver/integration_tests/doc.go
@@ -1,0 +1,4 @@
+// Package integration_tests require a running gitserver instance and are in
+// their own package to avoid cyclic dependencies between cmd/gitserver and
+// internal/gitserver
+package integration_tests
