+++ internal/gitserver/integration_tests/mocks.go
@@ -1,4 +1,4 @@
-package git
+package integration_tests
 
 import (
 	"io"
