+++ internal/gitserver/integration_tests/object_test.go
@@ -1,4 +1,4 @@
-package git
+package integration_tests
 
 import (
 	"context"
