+++ internal/gitserver/integration_tests/refs_test.go
@@ -1,4 +1,4 @@
-package git
+package integration_tests
 
 import (
 	"context"
